+++ lib/src/http_client.dart
@@ -18,6 +18,12 @@ class HttpClient {
         headers: headers, body: payload);
   }
 
+  Future<http.Response> doDelete(Uri uri,
+      {Map<String, String>? headers, Map<String, Object>? payload}) async {
+    return await _doCall(_HttpMethod.delete, uri,
+        headers: headers, body: payload);
+  }
+
   void _putDefaultHeaders(Map<String, String>? headers) {
     headers ??= {};
     headers['Accept'] = 'application/json';
@@ -41,8 +47,10 @@ class HttpClient {
         return await http.post(uri, headers: headers, body: jsonEncode(body));
       case _HttpMethod.put:
         return await http.put(uri, headers: headers, body: jsonEncode(body));
+      case _HttpMethod.delete:
+        return await http.delete(uri, headers: headers, body: jsonEncode(body));
     }
   }
 }
 
-enum _HttpMethod { get, post, put }
+enum _HttpMethod { get, post, put, delete }